flash attn fix [not supported for turing]

+++ openrlhf/trainer/ray/vllm_engine.py
@@ -232,7 +232,7 @@ def create_vllm_engines(
             "distributed_executor_backend": distributed_executor_backend,
             "max_model_len": max_model_len,
             "enable_prefix_caching": enable_prefix_caching,
-            "dtype": "bfloat16",
+            "dtype": "float16",
             "trust_remote_code": True,
             "full_determinism": full_determinism,
             "gpu_memory_utilization": gpu_memory_utilization,
